+++ src/main/java/baeksaitong/sofp/domain/profile/service/ProfileService.java
@@ -8,9 +8,9 @@ import baeksaitong.sofp.domain.profile.repository.ProfileRepository;
 import baeksaitong.sofp.domain.member.entity.Member;
 import baeksaitong.sofp.domain.profile.entity.Profile;
 import baeksaitong.sofp.global.error.exception.BusinessException;
-import baeksaitong.sofp.global.redis.RedisPrefix;
-import baeksaitong.sofp.global.redis.RedisService;
-import baeksaitong.sofp.global.s3.AwsS3Service;
+import baeksaitong.sofp.global.redis.constants.RedisPrefix;
+import baeksaitong.sofp.global.redis.service.RedisService;
+import baeksaitong.sofp.global.s3.service.AwsS3Service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
